Add comment continuation via on type on_type_formatting
One of my pet peeves is that when I am trying to write a comment block
on top of a method or class, hitting enter does not continue the comment
block but instead starts a new line for me. That is a huge hurdle for me
to write documentation comments.

On type formatting can be helpful here by continuing the comment block
and provide the indentation that the previous line of the comment block
message had.

This is especially helpful when one is writing code example inside a
comment block like:
```
  # This method acts in the following way:
  # ```
  # def foo
  #    puts "Hello World"
```
pressing `Enter` at this point should continue the comment block and
indent the next line of the comment block as much as the previous line:
```
  #   |
```