Make RuboCop runner use composition instead of inheritance
The current implementation of running RuboCop tasks is not composable.
Instead, we use inheritance to obtain and use the functionality of the
`RuboCop::Runner` class, which ends up coupling our interfaces too
tightly to that class.

Instead, this commit creates a `RuboCopRunner` class that is meant to
encapsulate the functionality of the `RuboCop::Runner` class. Both the
formatter and the diagnostic request classes are refactored to use this
new class in a composable way, and the implementations are much nicer as
a result.

This new architecture also allows us to add common functionality to how
RuboCop runs are made.