Don't parse expected/actual variables when printing diff
In the `assert_expectations` method, we already parse both `expected`
and `actual` json. So when they're passed to the `diff` method, the
`JSON.parse` will fail and aways fallback to the rescue block.

And because in the rescue block we print the objects directly, we only
get Hash result instead of prettified json content.