Upgrade bundler
We're seeing similar issues described in this issue:

https://github.com/dependabot/dependabot-core/issues/5917

And it looks like most of the issues came from bundler, not dependabot itself.

Given that

- ruby-lsp is using a relatively old version of bundler (2.3.7 was released in Feb 2022)
- dependabot now uses 2.3.25 by default

I think it's worth trying out a new bundler to see if it helps resolving the problem.

And if we compare this commit's change with https://github.com/Shopify/ruby-lsp/pull/410's,
we can see that the new bundler already generates the entries we'd like to see. So I think
the result is promising.