Start using `text_edit` instead of `insert_text` for completion items
With the current implementation when a completion request is received
with the require string like `"ruby_lsp/reque"`, the completion
suggestions are returned correctly, but when the user selects one of
the suggestions, for example, `ruby_lsp/requests/base_request`, then
their require string becomes `"ruby_lsp/sts/base_request"`.

Apparently, this happens because the contents of `insert_text` is open
to interpretation by the client:

> The `insertText` is subject to interpretation by the client side.
> Some tools might not take the string literally. For example
> VS Code when code complete is requested in this example
> `con<cursor position>` and a completion item with an `insertText` of
> `console` is provided it will only insert `sole`. Therefore it is
> recommended to use `textEdit` instead since it avoids additional client
> side interpretation.

This commit changes the implementation to use `text_edit` instead of
`insert_text` for completion items.