Update tests to be more resilient
We were relying on the structure of the `ruby_lsp` project for path
completions but that was a little too brittle.

This commit instead makes sure a temporary directory is created and a
known file structure is created within it. This directory is also added
to the load path so that the tests can successfully make completions.