Match .ruby-version behaviour of chruby-auto
The [`chruby-auto` script][0] in chruby will look for a `.ruby-version`
file by walking up the current directory tree. This is useful when
a project does not include a `.ruby-version` file, but the user has
configured a default version in their home directory, or in a parent
directory where they checkout all of their projects.

By default `chruby-auto` will walk up the directory tree [until it
hits a directory that doesn't exist][1]. When it reads a `.ruby-version`
in a directory, it only considers the version read from that file [if
it is not an empty string][2]. Otherwise, it continues walking up the
tree.

This commit matches that behaviour in our extension as well.

[0]: https://github.com/postmodern/chruby/blob/master/share/chruby/auto.sh
[1]: https://github.com/postmodern/chruby/blob/master/share/chruby/auto.sh#L6-L7
[2]: https://github.com/postmodern/chruby/blob/master/share/chruby/auto.sh#L9