Extract SyntaxTreeFormattingRunner
This largely mirrors the setup for RuboCopFormattingRunner and gives us
a place to put some SyntaxTree configuration code that we'd like to add
next.

While this new class's `#run` method only _needs_ a single `document`
argument, I chose to match the method signature of the RuboCop variant.
It feels like very little effort in service of paving the path for more
formatting runner classes. It might be worth defining this interface in
Sorbet, but I stopped short of doing that for now.