Use .streerc config when formatting w/ SyntaxTree
SyntaxTree supports a .streerc config file for specifying options like
print width and plugins. SyntaxTree picks these up for everything that
runs through its CLI. But since we're using SyntaxTree more directly, we
need to parse that file and pass it to SyntaxTree.format ourselves.

This is very similar to the setup used in SyntaxTree::CLI itself, and
as is he case with the RuboCop formatting runner, the use of a singleton
means we only parse that config file once.