Add run in terminal code lens
Because the test controller isn't able to handle dynamically generated
tests, the `Run` code lens is not able to run them. This means those tests
are not supported by any code lens anymore, while in prior versions they
could be run in the terminal.

This commits adds a new code lens `Run In Terminal` that will run the
tests in the terminal.