Add class name into code lens' test command
Currently, the `Run in Terminal` code lens always run all tests in the file
when it's clicked on a class definition. This means if a file contains
multiple test classes, all tests will be run when click on any of the class.

To resolve this, we need to specify the class name in the test command.
But when a class is provided, minitest would try to match the full class
name, including namespaces. This is not what we want, so we need to
quote `--name`'s value to make minitest do partial matching.