Migrate DocumentHighlight to Listener pattern
Since DocumentHighlight focuses on the type of the visited node instead
of the visiting event, I think it makes sense to introduce a new `on_node`
event for it. It saves us from expanding the nodes list in HighlightTarget
into a list of fragmented `on_<node>` events.

It should also benefit other requests that currently define the
`visit` method.