Push non-test class to the stack too
Currently we always pop a class from the stack when we run `after_class`,
but only push a test class when we run `on_class`. This means that if we
encounter a non-test class in the middle of a test class, we will mistakenly
pop the test class from the stack.

This commit fixes this by pushing a non-test class to the stack too. But
while doing so, we also need to check if `on_def` is in a test class or not.