Improve pipe completion if the user adds an accidental extra pipe
If we autocomplete pipes and the user still accidentally adds the second
one, we currently end up autocompleting in both cases and end up with 4
pipes.

This commits makes it so that if the user inserted a second pipe after
the autocomplete, we remove the pipe we added automatically and avoid
adding yet another one, resulting in two pipes.