Do not append extra pipe if the user adds a closing pipe manually
If the user adds a closing pipe manually, we currently end up autocompleting in both cases and end up with 3 pipes.

This commits makes it so that if the user inserted a closing pipe to the end of the block argument, we remove the pipe we added automatically and avoid adding yet another one, resulting in two pipes.