Fix cusor position when breaking after keyword
If the user breaks just after the keyword, this commit ensures the
cursor is placed back where the break happened. This is currently broken
in that the cursor will end up 1 character into the next line. There was
some discussion [1] about whether the cursor should remain where it
would normally be if the end insertion didn't happen, but it appears
moving it back to where the break happened is the intended behavior.

[1]: https://github.com/Shopify/ruby-lsp/pull/462#discussion_r1082642170