Start keeping a stack of namespaces instead of namespace names
This allows us to use the same stack for both the constant and method collection. For constants, keeping the current namespace on the stack allows us to refer to the current namespace name instantly, without having to build it from the stack entries. For methods, we need to keep the current namespace on the stack so that we can properly find the owner, i.e. the namespace that the method is defined in.