Choose other test framework when ruby-lsp-rails not present
ruby-lsp chooses 'rails' framework when rails is present in the gem
dependencies list, so that ruby-lsp-rails can add its own testing
commands. But this can be confusing if somebody installs ruby-lsp
without also installing ruby-lsp-rails on a rails project, because the
test commands will not show up at all. Ideally a dialogue in the
relevant extension would prompt the user to install ruby-lsp-rails, but
ruby-lsp should select the available testing framework.