Rename Request#response to Request#perform
As @vinistock pointed out, using `#response` seems to imply that the
request was already completed before it's called. In ruby-lsp's case,
many requests are used in this pattern:

```rb
FooRequest.new(args).response
```

In this context, it feels like the computation is done in
`FooRequest#initialize` and `#response` is just a getter, which is a bit
misleading.

So this commit renames `#response` to `#perform` to make it more clear.