Remove unnecessary hack from semantic highlighting
The hack was to prevent duplicated tokens being added to the constants
of a class declaration. However, this is not necessary as VS Code picks
up the right token even without it.

But the reason to delete it is that it's not covering all cases either.
For example, if the declared class has a namespace, like `class Foo::Bar`,
then it doesn't work at all.

Ideally, we should resolve the problem in a much better way, but for now
it's probably not worth the effort as it's not really causing any problems.