Bump Prism to 0.24 (#1385)
* Bump Prism to 0.24

* Update def_rescue fixture for newer Prism

Currently, the rescue folding range in the `def_rescue` test is actually
generated through the "pseudo" begin node's location, not the rescue node's.
This is because the rescue node's statement is empty, so we skip its range
in the folding range generation. Due to this, we actually are not testing
the folding range of the rescue node, but the "pseudo" begin block node's.

But after Prism 0.22, the "pseudo" begin block node's start position now
matches the def node's, so the test now fails.

So this commit updates the `def_rescue` fixture to actually test the
folding range of the rescue node, not the "pseudo" begin block node.