Supprot RuboCop's LSP mode
Resolves #1449.

This PR applies `RuboCop::LSP.enable`, which is the minimal API for RuboCop to handle in the context of LSP:
https://docs.rubocop.org/rubocop/1.61/usage/lsp.html#language-server-development

Furthermore, `RuboCop::LSP.enable` method and `RuboCop::LSP` module were introduced in RuboCop 1.61.0.
This PR ensures compatibility with existing versions of RuboCop dependencies.
However, in older versions of RuboCop prior to 1.60, `AutoCorrect: contextual` behaves the same as `AutoCorrect: true`.