Performance improvements for completions
Adds support for limiting the number of results from an index search,
and the `isIncomplete` flag for completion requets.

Completion requests will not return all matching results. This
dramatically improves editor performance.

In addition, limit the number of entries used to generate additional
details and documentation. This produces an additional performance
boost, especially when constants are defined in a very large number of
files. This can be the case for top-level modules in large codebases.