Add Rake document symbols (#1444)
* Add Rake document symbols

This commit adds document symbols for the `namespace` and `task` methods
in the Rake DSL.

* Pass URI to DocumentSymbol listener

This will allow us to enable certain document symbols based on the
filename. Specifically, we'll use it to only create document symbols for
namespace and task definitions in Rake files.

* fixup! Add Rake document symbols

* squash! Add Rake document symbols

We should only create document symbols for the `namespace` and `task`
DSL methods if the file is a Rakefile or a `.rake` file.

* squash! Add Rake document symbols

We should return early if the name is nil or empty. I've added tests for
an empty string, but it could also be an incomplete symbol.

* Extract node_name local var

---------

Co-authored-by: Andy Waite <andyw8@users.noreply.github.com>