Don't push group id to stack when encountering non-test classes (#1891)
Some tests use classes instead of modules for namespacing, such us:

```ruby
module Foo
  class Bar
    class MyTest < Minitest::Test
      def test_something
        assert true
      end
    end
  end
end
```

And when this happens, the current implementation would still push the
group id to the stack, which would cause the first test group (`MyTest`)
to already have non-nil group id. This will make the extension to think
there's another parent test group, which is not the case and will cause
no code lens to be shown.

This commit fixes this by not pushing the group id to the stack when
encountering non-test classes, and only pop the group id when the
encountered class is a test class.