Allow ENV to override GlobalState#test_library
In our project we have a very small amount of rspec tests and a large
amount of minitest tests. The test detection method prefers rspec over
minitest and as a result, the testing code lenses don't work for the
vast majority of our tests.

This change allows a user to set an ENV variable to force the
test_library to be what they choose.