Always push namespace to stack
Currently, the indexer doesn't index classes and modules that have dynamic
namespace. Similarly, those classes and modules' names are not pushed onto the
namespace stack. However, the stack is currently popped on every class or module
exit, which could lead to an incorrect stack popping.

This commit fixes the issue by always pushing the namespace to the stack, even
when the namespace is dynamic. This also means that we'll start indexing dynamic
namespaces' methods with the correct owner.