Improve performance of finding indexables
Currently, all folders and files in the current tree are turned into
IndexablePath, and then excluded files are filtered out after.

When there are large file trees that are meant to be excluded, this
results in a lot of unnecessary work.

ActiveStorage stores files in the `tmp` directory in many many small
folders. So does Bootsnap. Ruby LSP has to traverse all of these files,
even though the entire directory should just be ignored.

Rubocop has solved this by breaking the `includes` patterns up into many
patterns, applying the exclusions *before* the `Dir.glob`, so I followed
in their footsteps. This works great for exclusions that end in "**/*".
We still need to loop through all IndexablePath objects and see if
they're excluded, in the case that an extension was provided on the
excluded path, but this can cut down load time dramatically.

Before this PR in my Rails app, `indexables` took 76 seconds to run. Now
it takes, 0.17 seconds. Before and after code both return the same exact
file list.