Fix complex body indentation after end keyword
This commit fixes the auto indentation after the `end` keyword when the
body is more complex, such as when there are nested blocks.

Previously the following code:

```ruby
if foo
if bar
  baz
end
end
```

…would have been indented like this:

```rb
if foo
  if bar
  baz
end
end
```

Now it will be correctly indented like this:

```rb
if foo
  if bar
    baz
  end
end
```

It does so by indenting each line in the node location, instead of just
the first one.