Fix extra indentation after end keyword
This commit ensures auto-indentation after an `end` keyword does not add
extra indentation if the body is already indented.

Previously the following code:

```rb
if foo
  bar
end
```

…would have been indented like this:

```rb
if foo
    bar
end
```

Now it will be correctly indented like this:

```rb
if foo
  bar
end
```

It does so by finding the correct current indentation level, and
skipping indentation if it already matches.