Avoid overriding the start function of Client (#2163)
It turns out that the sendRequest function always calls the start
function of the client. So everything we put in the overriding function
will be called multiple times just during the initialization of the
extension.

While we currently only set the formatter and the server version in the
overriding function, which are fine when getting called multiple times,
it sets a dangerous precedent. For example, if we put another sendRequest
call in it, then it'd create an infinite loop.

Reference: https://github.com/microsoft/vscode-languageserver-node/blob/1ce790dd341a38da1fbf902cddc16ce456149de0/client/src/common/client.ts#L718