Add keywords to completion candidates (#2317)
* Bump Sorbet version so T.let on array constants is not required

* Add keywords to completion candidates

When typing `foo d`, VS Code would send a completion request for `d`.
And then when `foo do` is typed, it would do fuzzy search on the previous
completion candidates. So currently, the completion candidates for `foo do`
would be `define_singleton_method` as demonstrated in #2304.

Sorbet LSP solved this issue by adding keywords to the completion candidates.
And this PR does the same.