Correct capability objects for document symbol, folding range, and hover
ClientCapabilities are for the client (editor) to tell the server
(ruby-lsp) what it can do. The server should not be sending these back
to the client. Instead, the server should be sending back either boolean
values or Options objects.

This commit corrects the responses for the document symbol, folding
range, and hover requests to send back the correct values.