Document initialization options (#2464)
Document all initialization options

In a discussion on a recent PR[1], it was mentioned that it would be
helpful to document all of the initialization options. Something
noteworthy is that Neovim's LspConfig uses `init_options` rather than
the VSCode `initializationOptions`. Neovim converts `init_options` to
`initializationOptions` under the hood.

We can also roll the indexing section into the initialization options
section.

[1]: https://github.com/Shopify/ruby-lsp/pull/2445/files#r1722069148