Allow updating expectations when running tests
This commit adds a `WRITE_EXPECTATIONS` environment variable that allows
expectation snapshots to be automatically updated when running tests.
This is useful when making a change that affects many expectations which
would be painful to update manually. However, it should be used with
caution as it makes it easy to update expectations without really
thinking about it.