Capture issue with minitest spec code lenses
This commit adds a test to capture the current behavior of the minitest
spec code lenses. These are currently problematic because we sometimes
generate the same ID and command for different tests. Later commits will
fix both issues, but it would be good to have a test to see how the
output changes.