Use unique IDs for test code lenses
This commit ensures the IDs used for the code lenses are unique as far
as possible. This fixes an issue whereby multiple tests could have the
same ID and trying to run one of them would run a different one instead.

This could happen in a number of scenarios. For example, if you have
multiple test classes in the same file:

```
class FooTest < Minitest::Test
  def test_foo
  end
end

class BarTest < Minitest::Test
  def test_foo
  end
end
```

Or if you have nested specs:

```
describe Foo do
  describe "#bar" do
    it "returns true"
  end

  describe "#baz" do
    it "returns true"
  end
end
```

In both cases, the code lenses for the tests would have the same ID.

This commit fixes the issue by using a fully qualified selector for each
test, including the group stack and method name. This should ensure that
the IDs are unique.

For example, instead of `test_foo`, the ID would be `FooTest#test_foo`.
And instead of `returns true`, the ID would be `Foo#bar#returns true`.

The only cases that are not covered are dynamic module references which
can't be resolved statically. There is still a chance that two tests
have the same ID:

```
module abc::Foo
  class SomeTest < Minitest::Test
    def test_foo
    end
  end
end

module def::Foo
  class SomeTest < Minitest::Test
    def test_foo
    end
  end
end
```

Here both IDs would be `<dynamic_reference>::SomeTest#test_foo`. I'm not
sure how important it is to support this case—I expect it's quite rare.
One option would be to increment a counter and append it to the ID
somehow, but the added complexity doesn't seem worth it.

I'd prefer to replace the `name` argument with `id` and `label`, but
that would be a breaking change for plugins that are currently
referencing `add_test_code_lens`. As such, I added `id` as an optional
argument and handle it in the method. I've also kept backwards
compatibility between the extension and older versions of the LSP which
would not have the `id` field.