Use specific commands for minitest/spec tests
This commit ensures code lenses for minitest/spec tests have more
precise filter patterns so only the selected test is executed. In the
previous implementation, the filter pattern was too broad and would
potentially match multiple tests.

It works by generating a fully qualified selector for the test, matching
the format that is used internally by minitest. It increments a counter
for each group and prefixes the test name with the counter to ensure
uniqueness. This is potentially brittle as it makes an assumption about
the internal implementation of minitest, but it's the best we can do to
ensure only the selected test is run.

It's reasonably common to have multiple tests with the same name, for
example:

```
describe Foo do
  describe "#bar" do
    it "returns true"
  end

  describe "#baz" do
    it "returns true"
  end
end
```

Previously, clicking "Run In Terminal" for the `works` test under `#bar`
would run minitest with the following filter pattern:

> --name /works/

This would match multiple tests and run them both. With this change, the
filter pattern is more specific:

> --name "/^Foo#bar#test_0001_works$/"