Refactor client capabilities into its own class (#2805)
### Motivation

As pointed out by @andyw8, our implementation of the global state was growing a little too much and so was the body of `apply_options`. This PR extracts a new object to store client capabilities, so that we can organize concerns a bit better.

### Implementation

Extracted the client capabilities into an object and adapted the code that was using them for the new structure.

### Automated Tests

Updated existing tests.