Move `supports_progress` to global state (#2815)
### Motivation

We want to allow add-ons to show progress bars if the editor supports it. However, whether it supports it or not is currently a state of `Store`, which add-ons don't have access to. We need to put that in the global state.

### Implementation

Moved `supports_progress` to the global state.

### Automated Tests

I got rid of a test that wasn't very useful and that was difficult to avoid flakiness after this change.