Ensure we're handling client responses for window show message (#2814)
### Motivation

When the server makes a request to the client, the response doesn't include the requested method by default. We can make it so that add-ons need to include it as part of the message properties, but normally only the ID can be used to reconcile request and response.

We need to let add-ons include the method name as part of the properties so that we can delegate appropriately.

### Implementation

Our regular `process_message` would just ignore any responses because `message[:method]` is always `nil` in a response. We need to extract it from `result` and define that add-ons must include the `method` as part of it.

### Automated Tests

Adapted the test.