Add feature flag support for gradual rollouts (#2825)
### Motivation

Add the ability to rollout features to users based on percentages. This PR will essentially replace our experimental features configuration with a much more granular and controlled rolled out strategy.

Note: I'll leave removing the old experimental features setting to a future PR.

### Implementation

The idea is to combine the user's machine ID and the feature flag name to generate a number that can be checked against the desired rollout percentage.

With this strategy, we can guarantee that if we increase the percentage, users who already had the feature enabled will continue to do so.

From our side, the idea is to maintain a constant with the feature names and the desired rollout, so that we can gradually increase it as needed.

### Automated Tests

Added tests.