Add feature flag support to GlobalState (#2826)
### Motivation

This PR starts passing the enabled feature flags to the server and storing it in the global state. That way the Ruby LSP and its add-ons can check for specific feature flags. This also allows users of other editors to opt into features through `initializationOptions`.

### Implementation

- Started passing all enabled flag information to the server
- Started storing this information in the global state

### Automated Tests

Added unit tests to verify feature flag processing and retrieval functionality.