Ensure top level scripts are indexed (#2849)
### Motivation

With the optimizations to prevent descending into directories that we know are going to be excluded, we accidentally stopped indexing top level scripts.

### Implementation

Since the code now lists directories to eagerly exclude them, we need a separate path for the top level scripts.

### Automated Tests

Added a test that reproduces the issue.