Check target precision for call node in hover and definition (#2870)
### Motivation

Closes #2846

We had a regression in our call node target precisions. Essentially, the call node covers both the receiver and the message, but we only want to make the call node the target is the position covers the message location exactly.

### Implementation

Added another branch to the checks for whether the position is covered or not.

### Automated Tests

Added tests to ensure we don't regress.