Show multi-root workspace warning as progress instead (#2879)
### Motivation

We noticed that the warning for multi-root workspaces is sometimes ignored by users. The problem is that if you don't click anything, the LSP will never start because the promise is never resolved.

Let's show the warning under a timer and auto-dismiss so that the behaviour is less confusing.

### Implementation

I put the same warning inside a progress notification that auto-dismisses under 5 seconds.