Start returning composed bundle environment from initialize request (#2880)
### Motivation

Currently, we are always spawning the debug client with the incorrect environment. The reason is because the composed bundle process does a lot of heavy lifting to determine the exact environment variables that need to be set to apply an accurate Bundler configuration.

We need the debug client to have access to the composed environment if we want to be able to launch the debugger properly for all scenarios.

### Implementation

Started reading the bundle env and returning it as part of the initialize request.

### Automated Tests

Added a test to verify that we're returning the environment.