Allow indexing enhancements to create namespaces (#2857)
### Motivation

We realized that in order to support concerns `class_methods do...end`, indexing enhancements needed to be more powerful and have the liberty to create fake/temporary namespaces.

When trying to make that possible, we also realized that the API had many shortcomings that made enhancements harder than they had to be.

### Implementation

The idea is to instantiate enhancements with the declaration listener and then expose the API from it. That way, the declaration listener can hold important state like the code units cache and the current file path and enhancements only make flow adjustments and additions to the indexing process.

This PR also allows enhancements to create modules and classes, which allows us to support `class_methods do...end` in the Rails add-on and allows the RSpec add-on to support `let` and `subject` properly.

### Automated Tests

Added tests.